Added cart store and tests
