+++ src/store/item-store.js
@@ -1,10 +1,8 @@
 import { Store } from 'consus-flux';
 
 let item = null;
-let items = [];
 
 class ItemStore extends Store {
-
     getItem() {
         if (item === null) {
             return item;
@@ -14,11 +12,6 @@ class ItemStore extends Store {
             status: item.status
         };
     }
-
-    getAllItems(){
-        return items;
-    }
-
 }
 
 const store = new ItemStore();
@@ -28,17 +21,10 @@ store.registerHandler('ITEM_FOUND', data => {
         id: data.id,
         status: data.status
     };
-    items.push(item);
     store.emitChange();
 });
 
 store.registerHandler('NO_ITEM_FOUND', () => {
-    item = null;
-    store.emitChange();
-});
-
-store.registerHandler('CHECKOUT_SUCCESS', () => {
-    items = [];
     item = null;
     store.emitChange();
 });
